[run array] ci: customize ci , fix bugs in workflow
- ci: only trigger if changes
- doc,ci: add customize commit message to run ci
- wf: fix dtype bug

+++ src/BTVNanoCommissioning/workflows/ctag_emdileptt_valid_sf.py
@@ -344,8 +344,8 @@ class NanoProcessor(processor.ProcessorABC):
                     btagSFs(sjets, self.SF_map, weights, "DeepCSVB", syst_wei)
                     btagSFs(sjets, self.SF_map, weights, "DeepCSVC", syst_wei)
         else:
-            genflavor = ak.zeros_like(sjets.pt)
-            smflav = ak.zeros_like(smuon_jet.pt)
+            genflavor = ak.zeros_like(sjets.pt, dtype=int)
+            smflav = ak.zeros_like(smuon_jet.pt, dtype=int)
 
         # Systematics information
         if shift_name is None:
