+++ src/BTVNanoCommissioning/workflows/ttbar_validation.py
@@ -204,7 +204,7 @@ class NanoProcessor(processor.ProcessorABC):
                     btagSFs(sjets, self.SF_map, weights, "DeepCSVC", syst_wei)
 
         else:
-            genflavor = ak.zeros_like(sjets.pt)
+            genflavor = ak.zeros_like(sjets.pt, dtype=int)
 
         # Systematics information
         if shift_name is None:
